Swapped configs
